Finished implementing buffer scrolling
The buffer now scrolls vertically and horizontally when moving the
cursor to the appropriate edge of the screen. The point at which
scrolling begins can be adjusted with `editor->v_scroll_offset` and
`editor->h_scroll_offset`.

Implemented the Point structure to represent a position within the
Buffer, and added `editor_get_cursor_pos` to return a Point with the
current position of the cursor.

Also added comments explaining what each function does to the respective
header files.

+++ cursor.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 #include "cursor.h"
 
-Cursor *new_cursor(unsigned int row, unsigned int col) {
+Cursor *new_cursor(size_t row, size_t col) {
     Cursor *cursor = malloc(sizeof(Cursor));
     cursor->row = row;
     cursor->col = col;
